# expression: minus/plus/multiply (for Int and Real) check null iff overflow