--- a/expression/builtin_arithmetic_vec.go
+++ b/expression/builtin_arithmetic_vec.go
@@ -772,12 +772,12 @@ func (b *builtinArithmeticIntDivideIntSig) vecEvalInt(input *chunk.Chunk, result
 
 func (b *builtinArithmeticIntDivideIntSig) divideUU(result *chunk.Column, lhsI64s, rhsI64s, resultI64s []int64) error {
 	for i := 0; i < len(lhsI64s); i++ {
+		if result.IsNull(i) {
+			continue
+		}
 		lhs, rhs := lhsI64s[i], rhsI64s[i]
 
 		if rhs == 0 {
-			if result.IsNull(i) {
-				continue
-			}
 			if err := handleDivisionByZeroError(b.ctx); err != nil {
 				return err
 			}
@@ -793,11 +793,10 @@ func (b *builtinArithmeticIntDivideIntSig) divideUU(result *chunk.Column, lhsI64
 func (b *builtinArithmeticIntDivideIntSig) divideUS(result *chunk.Column, lhsI64s, rhsI64s, resultI64s []int64) error {
 	for i := 0; i < len(lhsI64s); i++ {
 		lhs, rhs := lhsI64s[i], rhsI64s[i]
-
+		if result.IsNull(i) {
+			continue
+		}
 		if rhs == 0 {
-			if result.IsNull(i) {
-				continue
-			}
 			if err := handleDivisionByZeroError(b.ctx); err != nil {
 				return err
 			}
@@ -805,9 +804,6 @@ func (b *builtinArithmeticIntDivideIntSig) divideUS(result *chunk.Column, lhsI64
 		} else {
 			val, err := types.DivUintWithInt(uint64(lhs), rhs)
 			if err != nil {
-				if result.IsNull(i) {
-					continue
-				}
 				return err
 			}
 			resultI64s[i] = int64(val)
@@ -818,12 +814,12 @@ func (b *builtinArithmeticIntDivideIntSig) divideUS(result *chunk.Column, lhsI64
 
 func (b *builtinArithmeticIntDivideIntSig) divideSU(result *chunk.Column, lhsI64s, rhsI64s, resultI64s []int64) error {
 	for i := 0; i < len(lhsI64s); i++ {
+		if result.IsNull(i) {
+			continue
+		}
 		lhs, rhs := lhsI64s[i], rhsI64s[i]
 
 		if rhs == 0 {
-			if result.IsNull(i) {
-				continue
-			}
 			if err := handleDivisionByZeroError(b.ctx); err != nil {
 				return err
 			}
@@ -831,9 +827,6 @@ func (b *builtinArithmeticIntDivideIntSig) divideSU(result *chunk.Column, lhsI64
 		} else {
 			val, err := types.DivIntWithUint(lhs, uint64(rhs))
 			if err != nil {
-				if result.IsNull(i) {
-					continue
-				}
 				return err
 			}
 			resultI64s[i] = int64(val)
@@ -844,12 +837,12 @@ func (b *builtinArithmeticIntDivideIntSig) divideSU(result *chunk.Column, lhsI64
 
 func (b *builtinArithmeticIntDivideIntSig) divideSS(result *chunk.Column, lhsI64s, rhsI64s, resultI64s []int64) error {
 	for i := 0; i < len(lhsI64s); i++ {
+		if result.IsNull(i) {
+			continue
+		}
 		lhs, rhs := lhsI64s[i], rhsI64s[i]
 
 		if rhs == 0 {
-			if result.IsNull(i) {
-				continue
-			}
 			if err := handleDivisionByZeroError(b.ctx); err != nil {
 				return err
 			}
@@ -857,9 +850,6 @@ func (b *builtinArithmeticIntDivideIntSig) divideSS(result *chunk.Column, lhsI64
 		} else {
 			val, err := types.DivInt64(lhs, rhs)
 			if err != nil {
-				if result.IsNull(i) {
-					continue
-				}
 				return err
 			}
 			resultI64s[i] = val