--- a/expression/builtin_arithmetic_vec.go
+++ b/expression/builtin_arithmetic_vec.go
@@ -792,10 +792,10 @@ func (b *builtinArithmeticIntDivideIntSig) divideUU(result *chunk.Column, lhsI64
 
 func (b *builtinArithmeticIntDivideIntSig) divideUS(result *chunk.Column, lhsI64s, rhsI64s, resultI64s []int64) error {
 	for i := 0; i < len(lhsI64s); i++ {
-		lhs, rhs := lhsI64s[i], rhsI64s[i]
 		if result.IsNull(i) {
 			continue
 		}
+		lhs, rhs := lhsI64s[i], rhsI64s[i]
 		if rhs == 0 {
 			if err := handleDivisionByZeroError(b.ctx); err != nil {
 				return err