@mmyj 2021-07-10
> No release note, Please follow https://github.com/pingcap/community/blob/master/contributors/release-note-checker.md

@BowenXiao1999 Please add a release note.