@mmyj 2021-07-11
Could you help to found all the vectorized expressions which can be optimized with ignoring check null,  and fix them in this PR?