@mmyj 2021-07-11
https://github.com/pingcap/tidb/pull/26109
This PR fix the `mod` expressions already.