@BowenXiao1999 2021-07-11
I have optimized all null check (not include mod) in this PR, you can view it commit by commit.

Will provide a benchmark here, but it may looks lengthy.